End day 14. Modified repositories. Modified services. Add image cards.

+++ BlackJack.DAL/Interfaces/ICardRepository.cs
@@ -4,6 +4,6 @@ namespace BlackJack.DAL.Interfaces
 {
     public interface ICardRepository 
     {
-        Card Get(int id);
+        Card Get(byte id);
     }
 }
